upbit_auto_trading/requirements.txt
# API 통신
requests==2.31.0
websockets==11.0.3
aiohttp==3.8.5

# 데이터 처리 및 분석
pandas==2.0.3
numpy==1.24.4
scipy==1.11.3
scikit-learn==1.3.0
ta-lib==0.4.28  # 기술적 분석 라이브러리

# 데이터베이스
sqlalchemy==2.0.20
pymysql==1.1.0
alembic==1.12.0  # 데이터베이스 마이그레이션

# 웹 인터페이스
flask==2.3.3
dash==2.13.0
plotly==5.16.1

# 설정 및 로깅
pyyaml==6.0.1
python-dotenv==1.0.0
loguru==0.7.0

# 테스트
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1

# 유틸리티
tqdm==4.66.1  # 진행 표시줄
schedule==1.2.0  # 작업 스케줄링
pytz==2023.3  # 시간대 관리